requirements.txt
Flask 
pymongo
pydantic
python-dotenv
PyJWT
Werkzeug